refactor(cmds): command names and structure reshaped
