+++ cmd/add.go
@@ -1,0 +1,42 @@
+/**
+ * Copyright © 2021-present Sveltin contributors <github@sveltin.io>
+ *
+ * Use of this source code is governed by Apache 2.0 license
+ * that can be found in the LICENSE file.
+ */
+
+// Package cmd ...
+package cmd
+
+import (
+	"fmt"
+
+	"github.com/spf13/cobra"
+)
+
+// addCmd represents the add command
+var addCmd = &cobra.Command{
+	Use:     "add",
+	Aliases: []string{"a"},
+	Short:   "Add content and metadata to a resource",
+	Long: `Used to add content and metadata to existing resources through its own subcommands.
+
+Run 'sveltin add -h' for further details.`,
+	Run: func(cmd *cobra.Command, args []string) {
+		fmt.Println("add called")
+	},
+}
+
+func init() {
+	rootCmd.AddCommand(addCmd)
+
+	// Here you will define your flags and configuration settings.
+
+	// Cobra supports Persistent Flags which will work for this command
+	// and all subcommands, e.g.:
+	// addCmd.PersistentFlags().String("foo", "", "A help for foo")
+
+	// Cobra supports local flags which will only run when this command
+	// is called directly, e.g.:
+	// addCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+}
