+++ cmd/newResource.go
@@ -30,7 +30,7 @@ import (
 
 var newResourceCmd = &cobra.Command{
 	Use:     "resource [name]",
-	Aliases: []string{"r"},
+	Aliases: []string{"r", "route"},
 	Short:   "Command to create a new resource",
 	Long: resources.GetASCIIArt() + `
 Command to create new resources.
